Slim down module by removing unnecessary options.
Remove option to select ASU header version and enforce use of v4.6 instead. The idea is that everytime a new ASU header version is available, a new version of ASU Brand module will be released. Also in this commit, an abstract class (AsuBrandBlockBase) was added to allow block classes (e.g. header and footer) to reuse code.